Restructure and organize files Also update README to include a photo of the draw command
