+++ cogs/Draw/colour.py
