+++ cogs/Draw/emoji.py
