+++ cogs/Draw/emoji_cache.py
