+++ cogs/Draw/errors.py
