+++ cogs/Draw/regexes.py
